refactor signup logic to component
